+++ pc/src/views/wshop/card_recharge/Api.ts
@@ -187,9 +187,12 @@ export async function findCount(
  * 根据 id 查找会员卡充值记录
  */
 export async function findById(
-  id: CardRechargeId,
+  id?: CardRechargeId,
   opt?: GqlOpt,
-) {
+): Promise<CardRechargeModel | undefined> {
+  if (!id) {
+    return;
+  }
   const data: {
     findByIdCardRecharge?: CardRechargeModel;
   } = await query({
@@ -210,12 +213,52 @@ export async function findById(
 }
 
 /**
+ * 根据 ids 查找会员卡充值记录
+ */
+export async function findByIds(
+  ids: CardRechargeId[],
+  opt?: GqlOpt,
+): Promise<CardRechargeModel[]> {
+  if (ids.length === 0) {
+    return [ ];
+  }
+  opt = opt || { };
+  opt.showErrMsg = false;
+  let models: CardRechargeModel[] = [ ];
+  try {
+    const data: {
+      findByIdsCardRecharge: CardRechargeModel[];
+    } = await query({
+      query: `
+        query($ids: [CardRechargeId!]!) {
+          findByIdsCardRecharge(ids: $ids) {
+            ${ cardRechargeQueryField }
+          }
+        }
+      `,
+      variables: {
+        ids,
+      },
+    }, opt);
+    models = data.findByIdsCardRecharge;
+  } catch (_err) { /* empty */ }
+  for (let i = 0; i < models.length; i++) {
+    const model = models[i];
+    await setLblById(model);
+  }
+  return models;
+}
+
+/**
  * 根据 ids 删除会员卡充值记录
  */
 export async function deleteByIds(
   ids: CardRechargeId[],
   opt?: GqlOpt,
-) {
+): Promise<number> {
+  if (ids.length === 0) {
+    return 0;
+  }
   const data: {
     deleteByIdsCardRecharge: Mutation["deleteByIdsCardRecharge"];
   } = await mutation({
@@ -238,7 +281,10 @@ export async function deleteByIds(
 export async function revertByIds(
   ids: CardRechargeId[],
   opt?: GqlOpt,
-) {
+): Promise<number> {
+  if (ids.length === 0) {
+    return 0;
+  }
   const data: {
     revertByIdsCardRecharge: Mutation["revertByIdsCardRecharge"];
   } = await mutation({
@@ -261,7 +307,10 @@ export async function revertByIds(
 export async function forceDeleteByIds(
   ids: CardRechargeId[],
   opt?: GqlOpt,
-) {
+): Promise<number> {
+  if (ids.length === 0) {
+    return 0;
+  }
   const data: {
     forceDeleteByIdsCardRecharge: Mutation["forceDeleteByIdsCardRecharge"];
   } = await mutation({
