+++ pc/src/views/wshop/pt/Api.ts
@@ -276,9 +276,12 @@ export async function updateById(
  * 根据 id 查找产品
  */
 export async function findById(
-  id: PtId,
+  id?: PtId,
   opt?: GqlOpt,
-) {
+): Promise<PtModel | undefined> {
+  if (!id) {
+    return;
+  }
   const data: {
     findByIdPt?: PtModel;
   } = await query({
@@ -299,12 +302,52 @@ export async function findById(
 }
 
 /**
+ * 根据 ids 查找产品
+ */
+export async function findByIds(
+  ids: PtId[],
+  opt?: GqlOpt,
+): Promise<PtModel[]> {
+  if (ids.length === 0) {
+    return [ ];
+  }
+  opt = opt || { };
+  opt.showErrMsg = false;
+  let models: PtModel[] = [ ];
+  try {
+    const data: {
+      findByIdsPt: PtModel[];
+    } = await query({
+      query: `
+        query($ids: [PtId!]!) {
+          findByIdsPt(ids: $ids) {
+            ${ ptQueryField }
+          }
+        }
+      `,
+      variables: {
+        ids,
+      },
+    }, opt);
+    models = data.findByIdsPt;
+  } catch (_err) { /* empty */ }
+  for (let i = 0; i < models.length; i++) {
+    const model = models[i];
+    await setLblById(model);
+  }
+  return models;
+}
+
+/**
  * 根据 ids 删除产品
  */
 export async function deleteByIds(
   ids: PtId[],
   opt?: GqlOpt,
-) {
+): Promise<number> {
+  if (ids.length === 0) {
+    return 0;
+  }
   const data: {
     deleteByIdsPt: Mutation["deleteByIdsPt"];
   } = await mutation({
@@ -328,7 +371,10 @@ export async function enableByIds(
   ids: PtId[],
   is_enabled: 0 | 1,
   opt?: GqlOpt,
-) {
+): Promise<number> {
+  if (ids.length === 0) {
+    return 0;
+  }
   const data: {
     enableByIdsPt: Mutation["enableByIdsPt"];
   } = await mutation({
@@ -353,7 +399,10 @@ export async function lockByIds(
   ids: PtId[],
   is_locked: 0 | 1,
   opt?: GqlOpt,
-) {
+): Promise<number> {
+  if (ids.length === 0) {
+    return 0;
+  }
   const data: {
     lockByIdsPt: Mutation["lockByIdsPt"];
   } = await mutation({
@@ -377,7 +426,10 @@ export async function lockByIds(
 export async function revertByIds(
   ids: PtId[],
   opt?: GqlOpt,
-) {
+): Promise<number> {
+  if (ids.length === 0) {
+    return 0;
+  }
   const data: {
     revertByIdsPt: Mutation["revertByIdsPt"];
   } = await mutation({
@@ -400,7 +452,10 @@ export async function revertByIds(
 export async function forceDeleteByIds(
   ids: PtId[],
   opt?: GqlOpt,
-) {
+): Promise<number> {
+  if (ids.length === 0) {
+    return 0;
+  }
   const data: {
     forceDeleteByIdsPt: Mutation["forceDeleteByIdsPt"];
   } = await mutation({
